backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
alembic==1.13.1
bcrypt==4.1.2
cryptography==42.0.5
python-dateutil==2.8.2
typing-extensions==4.9.0